+++ src/wavefunctions/mlccsd/mlccsd_class.F90
@@ -1193,7 +1193,7 @@ contains
    end subroutine cleanup_mlccsd
 !
 !
-   subroutine print_X1_diagnostics_mlccsd(wf, X, label)
+   subroutine print_X1_diagnostics_mlccsd(wf, X, n_amplitudes, label)
 !!
 !!    Print X1 diagnostics
 !!    Written by Sarai D. Folkestad, Nov 2019
@@ -1204,7 +1204,8 @@ contains
 !
       class(mlccsd), intent(in) :: wf
 !
-      real(dp), dimension(wf%n_es_amplitudes), intent(in) :: X
+      integer, intent(in) :: n_amplitudes
+      real(dp), dimension(n_amplitudes), intent(in) :: X
 !
       character(len=1), intent(in) :: label
 !
@@ -1214,7 +1215,7 @@ contains
 !
       integer :: a, i, ai, ai_full
 !
-      call wf%ccs%print_X1_diagnostics(X, label)
+      call wf%ccs%print_X1_diagnostics(X, n_amplitudes, label)
 !
       call mem%alloc(X_internal, wf%n_ccsd_v*wf%n_ccsd_o)
 !